reframe chain configs
