+++ src/chain-config.js
@@ -1,0 +1,400 @@
+
+export const ERC20List = [
+    // {symbol: 'MATIC', name: 'Matic Token', address: '0x0000000000000000000000000000000000001010', decimals: 18, icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/3890.png'},
+    {symbol: 'WMATIC', name: 'Wrapped Matic', address: '0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270', decimals: 18, icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/3890.png'},
+    {symbol: 'USDT', name: 'Tether USD', address: '0xc2132D05D31c914a87C6611C10748AEb04B58e8F', decimals: 6, icon: 'https://cdn.wherein.mobi/wormhole3/logo/usdt.png'},
+    {symbol: 'USDC', name: 'USD Coin', address: '0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174', decimals: 6, icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d/logo.png'},
+    {symbol: 'DAI', name: 'Dai Stablecoin', address: '0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063', decimals: 18, icon: 'https://polygonscan.com/token/images/mcdDai_32.png'},
+    {symbol: 'TEST-U', name: 'TEST USDT', address: '0x4cF89A27A27425d81C49c0B345e58A18De8A7273', decimals: 18, icon: 'https://cdn.wherein.mobi/wormhole3/logo/t-usdt.png'},
+    // {symbol: 'UNI', name: "Uniswap", address: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984', decimals: 18, icon: 'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984/logo.png'}
+]
+
+export const TokenIcon = {
+    'ETH': 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x2170Ed0880ac9A755fd29B2688956BD959F933F8/logo.png',
+    'BNB': 'https://s2.coinmarketcap.com/static/img/coins/64x64/1839.png',
+    'WBNB': 'https://s2.coinmarketcap.com/static/img/coins/64x64/1839.png',
+    'USDT': 'https://cdn.wherein.mobi/wormhole3/logo/usdt.png',
+    'TEST-U': 'https://cdn.wherein.mobi/wormhole3/logo/t-usdt.png',
+    'USDC': 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d/logo.png',
+    'UNI': 'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984/logo.png',
+    'BUSD': 'https://s2.coinmarketcap.com/static/img/coins/64x64/4687.png',
+    'MATIC': 'https://s2.coinmarketcap.com/static/img/coins/64x64/3890.png',
+    'WMATIC': 'https://s2.coinmarketcap.com/static/img/coins/64x64/3890.png',
+    'DAI': 'https://polygonscan.com/token/images/mcdDai_32.png',
+    'FRAX': 'https://cdn.wherein.mobi/wormhole3/logo/frax.png',
+    'NEAR': 'https://cdn.wherein.mobi/wormhole3/logo/near.png',
+    'AURORA': 'https://cdn.wherein.mobi/wormhole3/logo/aurora.png',
+    'GLMR': 'https://cdn.wherein.mobi/wormhole3/logo/moonbeam.png',
+    'WGLMR': 'https://cdn.wherein.mobi/wormhole3/logo/moonbeam.png',
+    'PNUT': 'https://cdn.wherein.mobi/nutbox/v2/1645760918196',
+    'NUT': "https://cdn.wherein.mobi/nutbox-v2/token/logo/nutcoin1.png",
+    'WIN': 'https://cdn.wherein.mobi/nutbox/v2/1650853172818',
+    'WNULS': 'https://cdn.wherein.mobi/wormhole3/logo/nuls-token.png',
+    'NULS': 'https://cdn.wherein.mobi/wormhole3/logo/nuls-token.png',
+    'USDTN': 'https://cdn.wherein.mobi/wormhole3/logo/usdt.png',
+    'USDCN': 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d/logo.png',
+    'AVAX': 'http://nassets.oss-us-west-1.aliyuncs.com/AVAX_9.png',
+    'HT': "http://nassets.oss-us-west-1.aliyuncs.com/HT_9.png",
+    'OKT': "http://nassets.oss-us-west-1.aliyuncs.com/OKT.png",
+    'ONE': "http://nassets.oss-us-west-1.aliyuncs.com/ONE_9.png",
+    'FTM': "http://nassets.oss-us-west-1.aliyuncs.com/FTM_9.png",
+    'KCS': "http://nassets.oss-us-west-1.aliyuncs.com/KCS_9.png",
+    'TRX': "http://nassets.oss-us-west-1.aliyuncs.com/TRX_9.png",
+    'IOTX': "http://nassets.oss-us-west-1.aliyuncs.com/IOTX_9.png",
+    'CRO': "http://nassets.oss-us-west-1.aliyuncs.com/CRO.png",
+    'METIS': "http://nassets.oss-us-west-1.aliyuncs.com/METIS.png",
+    'KLAY': "http://nassets.oss-us-west-1.aliyuncs.com/KLAY.png",
+    'NABOX': "http://nassets.oss-us-west-1.aliyuncs.com/NABOX_1.png",
+    'NVT': "http://nassets.oss-us-west-1.aliyuncs.com/NVT_9.png",
+    'ARB': "https://cdn.wherein.mobi/wormhole3/logo/ZjT_zBAS_400x400.png",
+    "MFERC": "https://cdn.wherein.mobi/community/logo/mobi-logo.jpg",
+    "ferc": 'https://cdn.wherein.mobi/nutbox/v2/1686657099126',
+    "wDMN": 'https://cdn.wherein.mobi/community/logo/damoon.jpg',
+    'wCH': 'https://cdn.wherein.mobi/community/logo/ch90.png',
+    'wPFP': 'https://cdn.wherein.mobi/community/logo/pfpdao.jpg',
+    "wAIR": 'https://cdn.wherein.mobi/community/logo/wh3-airdrop.png',
+    "BIBI": 'https://cdn.wherein.mobi/community/logo/bibi.png'
+}
+
+export const TokenName = {
+    'ETH': 'Ethereum',
+    'BNB': 'BSC Token',
+    'WBNB': 'Wrapped BNB',
+    'USDT': 'Tether USD',
+    'USDC': 'USD Coin',
+    'UNI': 'Uniswap',
+    'BUSD': 'BSC USD',
+    'MATIC': 'Polygon',
+    'TEST-U': 'TEST USDT',
+    'WMATIC': 'Wrapped Matic',
+    'DAI': 'Dai Stablecoin',
+    'FRAX': 'Frax',
+    'NEAR': 'Near',
+    'AURORA': 'Aurora',
+    'GLMR': 'GLMR',
+    'WGLMR': 'Wrapped GLMR',
+    'PNUT': 'Peanut',
+    'WNULS': 'Wrapped NULS',
+    'NULS': 'NULS',
+    'USDTN': 'USDTN',
+    'USDCN': 'USDCN',
+    'AVAX': 'AVAX',
+    'HT': 'HT',
+    'OKT': 'OKT',
+    'ONE': 'ONE',
+    'FTM': 'FTM',
+    'KCS': 'KCS',
+    'TRX': 'TRX',
+    'IOTX': 'IOTX',
+    'CRO': 'CRO',
+    'KAVA': 'KAVA',
+    'METIS': 'METIS',
+    'KLAY': 'KLAY',
+    'NABOX': 'Nabox_Token',
+    'NVT': 'NVT',
+    'ARB': 'Arbitrum',
+    "MFERC": 'mFERC',
+    "ferc": 'fair-ERC20',
+    "wCTLK": 'Wormhole Ctalks', 
+    "USDbC": 'USD Base Coin'
+}
+
+export const isWrappedToken = {
+    '0xbb4CdB9CBd36B01bD1cBaEBF2De08d9173bc095c': true,
+    '0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270': true,
+    '0x217dffF57E3b855803CE88a1374C90759Ea071bD': true,
+    '0xAcc15dC74880C9944775448304B263D191c6077F': true
+}
+
+export const EVM_CHAINS_ID = {
+    56: 'BNB Smart Chain',
+    137: 'Polygon',
+    42161: 'Arbitrum',
+    1284: 'Moonbeam',
+    119: 'ENULS',
+    59144: 'Linea',
+    8453: 'Base'
+}
+
+/**
+ * EVM Chains
+ */
+export const EVM_CHAINS = {
+    // ETH: {
+    //     rpc: "https://mainnet.infura.io/v3/0573c5d2e8c54ed79669b80cb3b78978",
+    //     scan: 'https://etherscan.io/',
+    //     id: 1,
+    //     main: {
+    //         name: 'Ethereum',
+    //         symbol: 'ETH',
+    //         icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x2170Ed0880ac9A755fd29B2688956BD959F933F8/logo.png',
+    //         decimals: 18
+    //     },
+    //     assets: {
+    //         USDT: {symbol: 'USDT', name: 'Tether USD', address: '0xdAC17F958D2ee523a2206206994597C13D831ec7', decimals: 6, icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x55d398326f99059fF775485246999027B3197955/logo.png'},
+    //         USDC: {symbol: 'USDC', name: 'USD Coin', address: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48', decimals: 6, icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d/logo.png'},
+    //         UNI:  {symbol: 'UNI', name: "Uniswap", address: '0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984', decimals: 18, icon: 'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984/logo.png'}
+    //     },
+    //     Multi_Config: {
+    //         rpcUrl: 'https://mainnet.infura.io/v3/0573c5d2e8c54ed79669b80cb3b78978',
+    //         multicallAddress: '0xeefba1e63905ef1d7acba5a8513c70307c1ce441',
+    //         interval: 3000,
+    //     }
+    // },
+    'BNB Smart Chain': {
+        // rpc:'https://bsc-dataseed.binance.org/',
+        rpc: 'https://bsc.publicnode.com',
+        scan: 'https://bscscan.com/',
+        tokenScan: 'https://pancakeswap.finance/info/tokens/',
+        chainName: 'BSC',
+        id: 56,
+        main: {
+            name: 'BNB',
+            symbol: 'BNB',
+            icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/1839.png',
+            decimals: 18
+        },
+        assets: {
+            // WETH: '0x2170Ed0880ac9A755fd29B2688956BD959F933F8',
+            WBNB: {symbol: 'WBNB', name: 'Wrapped BNB', address: '0xbb4CdB9CBd36B01bD1cBaEBF2De08d9173bc095c', decimals: 18, icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/1839.png'},
+            BUSD: {symbol: 'BUSD', name: 'BSC USD', address: '0xe9e7CEA3DedcA5984780Bafc599bD69ADd087D56', decimals: 18, icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/4687.png'},
+            USDT: {symbol: 'USDT', name: 'Tether USD', address: '0x55d398326f99059fF775485246999027B3197955', decimals: 18, icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x55d398326f99059fF775485246999027B3197955/logo.png'},
+            USDC: {symbol: 'USDC', name: 'USD Coin', address: '0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d', decimals: 18, icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d/logo.png'},
+            PNUT: {symbol: 'PNUT', name: 'Peanut', address: '0x705931A83C9b22fB29985f28Aee3337Aa10EFE11', decimals: 18}
+        },
+        walnutGraph: {
+            useTheGraph: true,
+            url: 'https://api.thegraph.com/subgraphs/name/terryyyyyy/walnutinbsc'
+        },
+        rewardCalculatorAddress: '0x6ab448C1C6e1870602d3FB867F167029bbFb3181',
+        Multi_Config: {
+            rpcUrl: 'https://bsc.publicnode.com',
+            multicallAddress: '0x41263cba59eb80dc200f3e2544eda4ed6a90e76c',
+            interval: 3000,
+        },
+        curation: '0x4695c30907E95A2153486D7a6372f3FBc2a89056',
+        popup: '0x7f50F5b0393E60D9c3cafdf652C1a900F5973AEB',
+        oracle: '0xfbD61B037C325b959c0F6A7e69D8f37770C2c550',
+        communityCuration: '0x6534b7A5a4dbF65b6DE92fF60dfB25f4Fdb7636B'
+    },
+    Polygon: {
+        rpc: 'https://polygon-rpc.com',
+        scan: 'https://polygonscan.com/',
+        chainName: 'Polygon',
+        tokenScan: 'https://app.uniswap.org/#/tokens/polygon/',
+        id: 137,
+        main: {
+            name: 'Polygon',
+            symbol: 'MATIC',
+            icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/3890.png',
+            decimals: 18
+        },
+        assets: {
+            WMATIC: {symbol: 'WMATIC', name: 'Wrapped Matic', address: '0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270', decimals: 18, icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/3890.png'},
+            USDT: {symbol: 'USDT', name: 'Tether USD', address: '0xc2132D05D31c914a87C6611C10748AEb04B58e8F', decimals: 6, icon: 'https://cdn.wherein.mobi/wormhole3/logo/usdt.png'},
+            USDC: {symbol: 'USDC', name: 'USD Coin', address: '0x2791Bca1f2de4661ED88A30C99A7a9449Aa84174', decimals: 6, icon: 'https://assets-cdn.trustwallet.com/blockchains/smartchain/assets/0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d/logo.png'},
+            DAI: {symbol: 'DAI', name: 'Dai Stablecoin', address: '0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063', decimals: 18, icon: 'https://polygonscan.com/token/images/mcdDai_32.png'},
+            // "TEST-U":{symbol: 'TEST-U', name: 'TEST USDT', address: '0x4cF89A27A27425d81C49c0B345e58A18De8A7273', decimals: 18, icon: 'https://cdn.wherein.mobi/wormhole3/logo/t-usdt.png'},
+        },
+        Multi_Config: {
+            rpcUrl: 'https://polygon-rpc.com',
+            multicallAddress: '0x11ce4B23bD875D7F5C6a31084f55fDe1e9A87507',
+            interval: 3000,
+        },
+        curation: '0x20eD66fD8da443128cd8E4E128BD02Da887dAe55',
+        popup: '0x9A6Dc03ceF711926155EFd010AaeF3BDD27be4f4',
+        oracle: '0x7F069df72b7A39bCE9806e3AfaF579E54D8CF2b9',
+        communityCuration: '0x90D0Ec60906f444567980E4688aC4525b7B63411'
+    },
+
+    Arbitrum: {
+        rpc: 'https://arb1.arbitrum.io/rpc',
+        scan: 'https://arbiscan.io/',
+        chainName: 'Arbitrum',
+        tokenScan: 'https://arbiscan.io/token/',
+        id: 42161,
+        main: {
+            name: 'ETH',
+            symbol: 'ETH',
+            decimals: 18,
+            icon: 'https://cdn.wherein.mobi/wormhole3/logo/ZjT_zBAS_400x400.png'
+        },
+        walnutGraph: {
+            useTheGraph: true,
+            url: 'https://api.thegraph.com/subgraphs/name/terryyyyyy/wanlutinarbitrum'
+        },
+        rewardCalculatorAddress: '0xa3e53F30C9cc6d174a98b311676e026535326f42',
+        assets: {
+          ARB: {symbol: 'ARB', name: 'Arbitrum', address: '0x912CE59144191C1204E64559FE8253a0e49E6548', decimals: 18},
+          USDC: {symbol: 'USDC', name: 'USD Coin', address: '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8', decimals: 6},
+          USDT: {symbol: 'USDT', name: 'Tether USD', address: '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9', decimals: 6},
+          DAI: {symbol: 'DAI', name: 'Dai Stablecoin', address: '0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1', decimals: 18},
+          ferc: {symbol: 'ferc', name: 'fair ERC20', address: '0xC365B8Cbde40cB902CaE1BDDf425a4c9E1f60d3f', decimals: 18},
+          MFERC: {symbol: 'MFERC', name: 'mFERC', address: '0xB4ee30dE6BF7e8F9eFBFcC9715021144DEFDe96F', decimals: 18},
+        },
+        Multi_Config: {
+            rpcUrl: 'https://arb1.arbitrum.io/rpc',
+            multicallAddress: '0x4B6EF997DbF23D320a6616ADf6499f4cdC00AB7c',
+            interval: 3000
+        },
+        curation: '0xb9D14987De32606572F856FE747F89E6e7f86B9B',
+        popup: '0x7d7d525e0937cA2B86c565F8fe56E4444BCc382D',
+        oracle: '0x735247fb0a604c0adC6cab38ACE16D0DbA31295F',
+        communityCuration: '0xDda7947F795B4636C68a14bC590fEb08cA69eef3'
+    },
+    Base: {
+        rpc: "https://mainnet.base.org/",
+        scan: "https://basescan.org/",
+        chainName: "Base",
+        tokenScan: "https://basescan.org/token/",
+        id: 8453,
+        main: {
+            name: 'ETH',
+            symbol: 'ETH',
+            decimals: 18,
+            icon: 'https://cdn.wherein.mobi/wormhole3/logo/base-icon.png'
+        },
+        walnutGraph: {
+            useTheGraph: true,
+            url: 'https://api.thegraph.com/subgraphs/name/terryyyyyy/walnutinbase'
+        },
+        rewardCalculatorAddress: "0xf6DDd65295Ca7A672C34043aa62f32C01FBfb29D",
+        assets: {
+            USDC: {symbol: 'USDC', name: 'USD Coin', address: '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913', decimals: 6},
+            USDbC: {symbol: 'USDbC', name: 'USD Base Coin', address: '0xd9aAEc86B65D86f6A7B5B1b0c42FFA531710b6CA', decimals: 6},
+        },
+        Multi_Config: {
+            rpcUrl: 'https://mainnet.base.org/',
+            multicallAddress: '0xcA11bde05977b3631167028862bE2a173976CA11',
+            interval: 3000
+        },
+        curation: '',
+        popup: '',
+        oracle: '0x52cbE0f49CcdD4Dc6E9C13BAb024EABD2842045B',
+        communityCuration: "0x3686218f11c58ca46479acc5DdEE7e41374dF73A"
+    },
+    Linea: {
+        rpc: 'https://rpc.linea.build',
+        scan: 'https://lineascan.build/',
+        chainName: 'Linea',
+        tokenScan: 'https://lineascan.build/token/',
+        id: 59144,
+        main: {
+            name: 'ETH',
+            symbol: 'ETH',
+            decimals: 18,
+            icon: 'https://l2beat.com/icons/linea.png'
+        },
+        walnutGraph: {
+            useTheGraph: false,
+            url: 'https://linea-graph.nutbox.app/v1/common/search'
+        },
+        assets: {
+            BUSD: {symbol: 'BUSD', name: 'BSC USD', address: '0x7d43AABC515C356145049227CeE54B608342c0ad', decimals: 18, icon: 'https://s2.coinmarketcap.com/static/img/coins/64x64/4687.png'},
+        },
+        Multi_Config: {
+            rpcUrl: 'https://rpc.linea.build',
+            multicallAddress: '0xcA11bde05977b3631167028862bE2a173976CA11',
+            interval: 3000
+        },
+        rewardCalculatorAddress: '0xF21649D901A082772Bd7B5d5eD5039C7a43A5789',
+        curation: '0x7d7d525e0937cA2B86c565F8fe56E4444BCc382D',
+        popup: '0xb9D14987De32606572F856FE747F89E6e7f86B9B',
+        oracle: null,
+        communityCuration: '0xdf74187fe7de3C6F6b2f01fF76E021e3b470eDb0'
+    },
+    'ENULS': {
+        rpc: 'https://evmapi.nuls.io',
+        scan: 'https://evmscan.nuls.io',
+        chainName: 'ENULS',
+        tokenScan: 'https://evmscan.nuls.io/token/',
+        id: 119,
+        main: {
+            name: 'NULS',
+            symbol: 'NULS',
+            decimals: 18,
+            icon: 'https://cdn.wherein.mobi/wormhole3/logo/ENULS.png'
+        },
+        assets: {
+            WNULS: {symbol: 'WNULS', name: 'Wrapped NULS', address: '0x217dffF57E3b855803CE88a1374C90759Ea071bD',decimals: 18, icon: 'https://cdn.wherein.mobi/wormhole3/logo/nuls-token.png'},
+
+            USDTN: {symbol: 'USDTN', name: 'USDTN', address:  '0x9e5d124cd49671f3f7b54d4aef43b3930bcf6de7', decimals: 18, icon: ''},
+
+            USDCN: {symbol: 'USDCN', name:'USDCN', address: '0xb13bb925d62adc0ea0da95f70e7f7a09effd4f9e', decimals: 18, icon: ''},
+
+            ETH: {symbol: 'ETH', name: 'ETH', address: '0xf573632c60de84f30486555e5199911fbd00835d', decimals: 18, icon: ''},
+
+            BNB: {symbol: 'BNB', name: 'BNB', address: '0x70a07c8dce789d9cc249b3f816bdacd2bbed66cc', decimals: 18, icon: ''},
+
+            AVAX: {symbol: 'AVAX', name: 'AVAX', address: '0x275f8b996dca35bca02cf2edb0f883d83decabf0', decimals: 18, icon: ''},
+
+            HT: {symbol: 'HT', name: 'HT', address: '0x1853f95c1e21d6f43a667dd01f174bdf7828d4af', decimals: 18, icon: ''},
+
+            OKT: {symbol: 'OKT', name: 'OKT', address: '0x3a5b9a4fd8fcac2cf2b09bb69f15f0c4a2a34b10', decimals: 18, icon: ''},
+
+            ONE: {symbol: 'ONE', name: 'ONE', address: '0x6770b3ada065fe3e33d93de76ad22ead07d8c8d5', decimals: 18, icon: ''},
+
+            MATIC: {symbol: 'MATIC', name: 'MATIC', address: '0x39f36e05630ac18cd4cbc4a787f58fa280066208', decimals: 18, icon: ''},
+
+            FTM: {symbol: 'FTM', name: 'FTM', address: '0x21cd8e25a5504c55c26d0c2b84adffd3607e6d17', decimals: 18, icon: ''},
+
+            KCS: {symbol: 'KCS', name: 'KCS', address: '0x5ec1c7d67ab36532ecafa8464717a7b282a909b3', decimals: 18, icon: ''},
+
+            TRX: {symbol: 'TRX', name: 'TRX', address: '0x6949e317f000e07496654dc68fe252739ffcbc10', decimals: 6, icon: ''},
+
+            IOTX: {symbol: 'IOTX', name: 'IOTX', address: '0x3e1c9a9e67d00fc55c5e2be28ea00dadc15cc065', decimals: 18, icon: ''},
+
+            CRO: {symbol: 'CRO', name: 'CRO', address: '0x4ced4e629a0a314d5bae83f93edf80df5c2f13a1', decimals: 18, icon: ''},
+
+            KAVA: {symbol: 'KAVA', name: 'KAVA', address: '0xc27956fd894d7192892bfe8d215a5b25dc893da0', decimals: 18, icon: ''},
+
+            METIS: {symbol: 'METIS', name: 'METIS', address: '0x15be367243896393c097660e9c7a693ff3d9dedd', decimals: 18, icon: ''},
+
+            KLAY: {symbol: 'KLAY', name: 'KLAY', address: '0x518111b3d78a11ff076f23f24fda1fae93f1c32f', decimals: 18, icon: ''},
+
+            NABOX: {symbol: 'NABOX', name: 'Nabox_Token', address: '0xd9bea2dff06366c84ac2654308c5be9cb8750889', decimals: 18, icon: ''},
+
+            NVT: {symbol: 'NVT', name: 'NVT', address: '0xCBE315d0163E372A82436F67C64097d6634394a9', decimals: 8, icon: ''},
+        },
+        walnutGraph: {
+            useTheGraph: false,
+            url: 'https://enuls-graph.nutbox.app/v1/common/search'
+        },
+        rewardCalculatorAddress: '0xb035C2bD1F0D2f36d78B9161Ca8F04aA81DCD20E',
+        Multi_Config: {
+            rpcUrl: 'https://evmapi.nuls.io',
+            multicallAddress: '0x383870Ae4E834155192cEce2fb5B0528CE0790E9',
+            interval: 3000
+        },
+        curation: '0xb9D14987De32606572F856FE747F89E6e7f86B9B',
+        popup: '0xA55deCFdcc7B57f67f4E60DfF6a8226A4cEcC268',
+        communityCuration: '0x1Ac88fa4ec4923835b25b21cE2061b25e0A4b05d'
+    },
+    Moonbeam: {
+        rpc: 'https://rpc.api.moonbeam.network',
+        scan: 'https://moonbeam.moonscan.io/',
+        tokenScan: 'https://moonbeam.moonscan.io/token/',
+        chainName: 'MoonBeam',
+        id: 1284,
+        main: {
+            name: 'GLMR',
+            symbol: 'GLMR',
+            decimals: 18,
+            icon: 'https://cdn.wherein.mobi/wormhole3/logo/moonbeam.png'
+        },
+        assets: {
+          WGLMR: {symbol: 'WGLMR', name: 'Wrapped GLMR', address: '0xAcc15dC74880C9944775448304B263D191c6077F', decimals: 18},
+          USDC: {symbol: 'USDC', name: 'USD Coin', address: '0x818ec0A7Fe18Ff94269904fCED6AE3DaE6d6dC0b', decimals: 6},
+          USDT: {symbol: 'USDT', name: 'Tether USD', address: '0xeFAeeE334F0Fd1712f9a8cc375f427D9Cdd40d73', decimals: 6},
+          DAI: {symbol: 'DAI', name: 'Dai Stablecoin', address: '0x765277EebeCA2e31912C9946eAe1021199B39C61', decimals: 18},
+        },
+        Multi_Config: {
+            rpcUrl: 'https://rpc.api.moonbeam.network',
+            multicallAddress: '0xc3035058E5c81349d881267DeC610B4D4A983a76',
+            interval: 3000
+        },
+        curation: '0x38A096b9cF4Cfb881BaB778Ee8B5437d2A412214',
+        popup: '0xF3d821614AD0ECC0AcA83D4b137D2Fde7de51434',
+        communityCuration: '0xA7E27609a0185721e9CFda1188c6797A6d28f452'
+    }
+}
