+++ src/components/ReferralRewardItem.vue
@@ -102,7 +102,8 @@
 </template>
 
 <script>
-import {TokenIcon, errCode} from "@/config";
+import { errCode } from "@/config";
+import { TokenIcon } from "@/chain-config";
 import Avatar from "@/components/Avatar";
 import defaultAvatar from "@/assets/icon-default-avatar.svg";
 import ChainTokenIcon from "@/components/ChainTokenIcon";
@@ -110,7 +111,7 @@ import {useWindowSize} from "@vant/use";
 import {parseTimestamp, formatAmount, formatPrice} from "@/utils/helper";
 import { formatAddress } from '@/utils/tool'
 import { mapGetters, mapState } from 'vuex'
-import { EVM_CHAINS, EVM_CHAINS_ID } from "@/config";
+import { EVM_CHAINS, EVM_CHAINS_ID } from "@/chain-config";
 import { getPriceFromOracle } from '@/utils/asset'
 import RewardHistoryList from "@/components/RewardHistoryList";
 import {accountChanged, getAccounts} from "@/utils/web3/account";
