+++ src/views/community/CommunityCredit.vue
@@ -62,7 +62,7 @@ import { getCommunityMembers } from '@/api/api'
 import { sleep, formatAmount, formatPrice, parseTimestamp } from '@/utils/helper'
 import { notify } from '@/utils/notify'
 import { getTokenBalancesOfUsers } from '@/utils/asset'
-import { EVM_CHAINS, EVM_CHAINS_ID } from '@/config'
+import { EVM_CHAINS, EVM_CHAINS_ID } from '@/chain-config'
 import Avatar from "@/components/Avatar.vue";
 
 export default {
