+++ src/views/user/ReferralRewardView.vue
@@ -37,14 +37,14 @@
 import { mapGetters, mapState } from 'vuex'
 import {notify} from "@/utils/notify";
 import {formatAddress} from "@/utils/tool";
-import { EVM_CHAINS, EVM_CHAINS_ID } from '@/config';
+import { EVM_CHAINS, EVM_CHAINS_ID } from '@/chain-config';
 import { formatAmount } from '@/utils/helper'
 import {accountChanged, getAccounts} from "@/utils/web3/account";
 import { setupNetwork } from '@/utils/web3/web3'
 import { setCurationIsFeed,
   getCommunityPendingRewards, getCommunityAuthorPendingRewards, setCommunityRewardClaimed,
   setCommunityAuthorRewardClaimed, getPendingClaimSummary } from '@/api/api'
-import {TokenIcon} from "@/config";
+import {TokenIcon} from "@/chain-config";
 import {useWindowSize} from "@vant/use";
 import ReferralRewardItem from "@/components/ReferralRewardItem";
