+++ src/views/user/RewardPromotion.vue
@@ -112,7 +112,8 @@
 <script>
 import RewardHistoryList from "@/components/RewardHistoryList";
 import {useWindowSize} from "@vant/use";
-import {EVM_CHAINS, EVM_CHAINS_ID, TokenIcon} from "@/config";
+import {EVM_CHAINS, EVM_CHAINS_ID} from "@/chain-config";
+import { TokenIcon } from "@/chain-config";
 import {mapGetters, mapState} from "vuex";
 import {formatAddress} from "@/utils/tool";
 import {formatAmount} from "@/utils/helper";
